- Added screen scaling to any resolution. - Cleaned up event system. - Updated version number. - Updated main menu. - Fixed docs.
git-svn-id: https://openxcom.svn.sourceforge.net/svnroot/openxcom/trunk@178 11c4a5ed-7179-4546-81a6-367b3be9d812

+++ src/Engine/Palette.h
@@ -58,7 +58,7 @@ public:
 	 * @param block Requested block.
 	 * @return Color position.
 	 */
-	static inline int blockOffset(int block) { return block*16; }
+	static inline Uint8 blockOffset(Uint8 block) { return block*16; }
 	/// Position of the background colors block in an X-Com palette (used for background images in screens).
 	static const int backPos = 224;
 };
